rename: tokens start with _ for partial
